backend/requirements.txt
fastapi
uvicorn
openai
transformers
torch
faiss-cpu
emoji
pandas
kagglehub
numpy
requests
matplotlib